+++ src/control_plane/migration/rabbitmq/mod.rs
@@ -1,5 +1,9 @@
 mod backup_rabbitmq_vhost;
 mod rabbitmq_backup_options;
+mod rabbitmq_restore_options;
+mod restore_rabbitmq_vhost;
 
 pub(crate) use backup_rabbitmq_vhost::backup_rabbitmq_vhost;
 pub(crate) use rabbitmq_backup_options::RabbitMqBackupOptions;
+pub(crate) use rabbitmq_restore_options::RabbitMqRestoreOptions;
+pub(crate) use restore_rabbitmq_vhost::restore_rabbitmq_vhost;